requirements.txt
openai>=1.10.0
langchain>=0.2.0
langchain-community>=0.0.30
langchain-openai>=0.0.7
faiss-cpu>=1.7.4
streamlit>=1.35.0
pandas>=2.0.0
scikit-learn>=1.4.0
plotly>=5.20.0
python-dotenv>=1.0.0
tiktoken>=0.5.1
